requirements.txt
transformers[torch]==4.44.2
numpy==1.23.5
scikit-learn==1.6.1
tqdm==4.67.1
googletrans==4.0.2
langdetect==1.0.9
datasets==2.9.0
evaluate==0.4.0
matplotlib==3.7.1
pandas==1.5.3
torch>=1.11.0
tiktoken==0.8.0
protobuf==5.29.3
gdown==5.2.0
sentencepiece==0.2.0
accelerate==1.3.0
wandb==0.19.4